% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_getApproxCovar.R
\name{binStats}
\alias{binStats}
\title{Bin the test statistic into equally sized bins}
\usage{
binStats(z, nBins = 82L, binEdges = c(-4.1, 4.1))
}
\arguments{
\item{z}{the matrix of permuted test statistics}

\item{nBins}{an integer, the number of bins}

\item{binEdges}{A vector of length 2 with the outer bin edges}
}
\value{
Matrix of binned test statistics
}
\description{
Bin the test statistic into equally sized bins
}
