% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_calcWeights.R
\name{calcWeights}
\alias{calcWeights}
\title{Obtain weights as posterior probabilities to calculate the consensus null}
\usage{
calcWeights(logDensPerm, fdr)
}
\arguments{
\item{logDensPerm}{A matrix with B rows  of logged density estimates of the B
permutation distributions, and p columns for the p observed test
statistics}

\item{fdr}{A vector of local false discovery rates for the observed tests
statistics of length p}
}
\value{
A vector of weights of length B
}
\description{
Obtain weights as posterior probabilities to calculate the consensus null
}
