% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_estP0.R
\name{estP0}
\alias{estP0}
\title{Estimate the fraction of true null hypotheses.}
\usage{
estP0(statObs, fitAll, z0quantRange, smooth.df, evalVal)
}
\arguments{
\item{statObs}{A vector of observed z-values}

\item{fitAll}{the estimated normal null}

\item{z0quantRange}{a number of quantiles between 0 and 0.5}

\item{smooth.df}{degrees of freedom for the spline smoother}

\item{evalVal}{the value of q at which to evaluate the spline}
}
\value{
The estimated null fraction, the value of the spline evaluated
 at the first element of z0quantRange
}
\description{
Estimate the fraction of true null hypotheses.
}
\details{
A natural spline is used over a range of intervals.
Based on the qvalue::qvalue() function and
Storey and Tibshirani, 2003
}
