% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_getCperm.R
\name{getCperm}
\alias{getCperm}
\title{Obtain a null covariance matrix of binned test statistics}
\usage{
getCperm(zValsMat, nBins = 82L, binEdges = c(-4.1, 4.1))
}
\arguments{
\item{zValsMat}{The pxB matrix of permutation z-values in the columns}

\item{nBins}{an integer, the number of bins}

\item{binEdges}{A vector of length 2 with the outer bin edges}
}
\value{
The covariance matrix of binned z-values
}
\description{
Obtain a null covariance matrix of binned test statistics
}
\note{
This is not the covariance matrix of the p test statistic, nor of the
   data! It is an approximate covariance matrix of binned test statistics for
   visualization purposes.
}
