% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_getFdr.R
\name{getFdr}
\alias{getFdr}
\title{Calculate tail-area (Fdr) and local (fdr) false discovery rates,
   based on a certain null distribution}
\usage{
getFdr(zValObs, G0, g0, zSeq, fdr, p, p0, ...)
}
\arguments{
\item{zValObs}{Vector of observed z-values}

\item{G0}{null distribution function}

\item{g0}{Null density function}

\item{zSeq}{Support of the density estimation}

\item{p}{the number of hypotheses}

\item{p0}{The estimated fraction of null hypotheses}

\item{...}{more arguments, ignored}
}
\value{
A list with components
\item{Fdr}{Tail are false discovery rate}
\item{fdr}{Local false discovery rate}
\item{p0}{The proportion of true null hypotheses}
}
\description{
Calculate tail-area (Fdr) and local (fdr) false discovery rates,
   based on a certain null distribution
}
