% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_getTstat.R
\name{getTstat}
\alias{getTstat}
\title{A function to obtain a t-test statistic efficiently. For internal use only}
\usage{
getTstat(y1, y2, mm, nn)
}
\arguments{
\item{y1, y2}{vectors of obsereved values in the two groups}

\item{mm, nn}{number of observations in the corresponding groups}
}
\value{
A list with items
\item{tstat}{The t-test statistic}
\item{df}{The degrees of freedom (Welch approximation)}
}
\description{
A function to obtain a t-test statistic efficiently. For internal use only
}
