% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_interpKDE.R
\name{interpKDE}
\alias{interpKDE}
\title{Evaluate a kernel density through interpolation}
\usage{
interpKDE(kdeObj, newData)
}
\arguments{
\item{kdeObj}{the kernel density fit}

\item{newData}{A vector of values at which there is to be interpolated}
}
\value{
the interpolated value vector of the length of newData
}
\description{
Evaluate a kernel density through interpolation
}
