% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_plotNull.R
\name{plotNull}
\alias{plotNull}
\title{Plot the obtained null distribution along with a histogram of observed test
statistics}
\usage{
plotNull(fit)
}
\arguments{
\item{fit}{an object returned by the fdrCorrect() (or testDAA()) function}
}
\value{
a ggplot2 plot object
}
\description{
Plot the obtained null distribution along with a histogram of observed test
statistics
}
