% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_pt.edit.R
\name{pt.edit}
\alias{pt.edit}
\title{A custom function to calculate the distribution function of the t-test
 statistic. For internal use only}
\usage{
pt.edit(q)
}
\arguments{
\item{q}{a vector with t-statistic and degrees of freedom}
}
\value{
A value between 0 and 1, the evaluation of the cdf
}
\description{
A custom function to calculate the distribution function of the t-test
 statistic. For internal use only
}
