% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_ptEdit.R
\name{ptEdit}
\alias{ptEdit}
\title{A custom function to calculate the distribution function of the t-test
 statistic. For internal use only}
\usage{
ptEdit(q)
}
\arguments{
\item{q}{a vector with t-statistic and degrees of freedom}
}
\value{
A value between 0 and 1, the evaluation of the cdf
}
\description{
A custom function to calculate the distribution function of the t-test
 statistic. For internal use only
}
