% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_qnormIf.R
\name{qnormIf}
\alias{qnormIf}
\title{Convert to z-values or not}
\usage{
qnormIf(zValues, quantileFun, testPargs, stats)
}
\arguments{
\item{zValues}{a boolean, is conversion to z-values required}

\item{quantileFun}{the quantile function}

\item{testPargs}{additional arguments to the quantile functions}

\item{stats}{the vector of test statistics}
}
\value{
z-values or original statistics
}
\description{
Convert to z-values or not
}
