% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_ptEdit.R
\name{qtEdit}
\alias{qtEdit}
\title{A custom function to calculate the quantile function of the t-test
 statistic. For internal use only}
\usage{
qtEdit(p)
}
\arguments{
\item{p}{a vector with quantile and degrees of freedom}
}
\value{
the corresponding quantile
}
\description{
A custom function to calculate the quantile function of the t-test
 statistic. For internal use only
}
