% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_quantCorrect.R
\name{quantCorrect}
\alias{quantCorrect}
\title{Correct quantiles by not returning 0 or 1}
\usage{
quantCorrect(quants)
}
\arguments{
\item{quants}{A vector of quantiles}
}
\value{
The same vector of quantiles but without 0 or 1 values
}
\description{
Correct quantiles by not returning 0 or 1
}
