% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_stabExp.R
\name{stabExp}
\alias{stabExp}
\title{A function to numerically stabilize an exponentiation. For internal use only}
\usage{
stabExp(exps)
}
\arguments{
\item{exps}{the vector to be exponentiated}
}
\value{
the vector with the maximum subtracted
}
\description{
A function to numerically stabilize an exponentiation. For internal use only
}
