% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_testDAA.R
\docType{methods}
\name{testDAA}
\alias{testDAA}
\alias{testDAA,phyloseq-method}
\alias{testDAA,matrix-method}
\title{A function to test for differential absolute abundance on a phyloseq object}
\usage{
testDAA(Y, ...)

\S4method{testDAA}{phyloseq}(Y, groupName, FCname, ...)

\S4method{testDAA}{matrix}(Y, FC, x, S = rowSums(Y), tieBreak = "none",
  ...)
}
\arguments{
\item{Y}{a n-by-p sequence count tabel with taxa in the columns}

\item{...}{passed on to the testDAA() function}

\item{groupName}{A character string, the name of a variable in physeq
indicating the grouping factor}

\item{FCname}{A character string, the name of a variable in physeq
containing the total cell count}

\item{FC}{a vector of length n with total flow cytometry cell counts}

\item{x}{a grouping factor of length n}

\item{S}{a vector of library sizes. Will be calculated f}

\item{tieBreak}{A character string, specifying the tie break doctrine to use.}

\item{physeq}{A phyloseq object}

\item{...}{other arguments passed on to the fdrCorrect function}
}
\value{
See the fdrCorrect() function

see ?fdrCorrect()
}
\description{
A function to test for differential absolute abundance on a phyloseq object

Perform the complete differential absolute abundance analysis
}
\details{
This function breaks ties, and filters empty rows and columns
prior to statistical testing
}
\examples{
testVanDePutte = testDAA(Vandeputte, "Health.status", "absCountFrozen")
testMat = testDAA(as.matrix(otu_table(Vandeputte)),
get_variable(Vandeputte, "Health.status"),
get_variable(Vandeputte,"absCountFrozen"))
}
