% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_tiebreak.R
\name{tiebreak}
\alias{tiebreak}
\title{A function to break the ties because of zeroes in the sequencing counts,
and prepare the final matrix}
\usage{
tiebreak(Y, FC, S = rowSums(Y), tieBreak = "none", tol = 1e-04,
  maxIter = 1L)
}
\arguments{
\item{Y}{the sequence count matrix}

\item{FC}{the vector of total cell counts}

\item{S}{the library sizes. By default calculated from Y}

\item{tieBreak}{the tiebreaking protocol}

\item{tol}{a convergence tolerance}

\item{maxIter}{Maximum number of iterations}
}
\value{
The matrix to analyze
}
\description{
A function to break the ties because of zeroes in the sequencing counts,
and prepare the final matrix
}
