% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_data.R
\docType{data}
\name{Vandeputte}
\alias{Vandeputte}
\title{Microbiomes of Crohn's disease patients and healthy controls}
\format{
A phyloseq object with an OTU-table and sample data
\describe{
  \item{otu_table}{Count data matrix of 234 taxa in 135 samples}
  \item{sample_data}{Data frame of patient covariates}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/29143816}
}
\usage{
Vandeputte
}
\description{
Microbiome sequencing data of Crohn's disease patients,
and healthy controls, together with other baseline covariates.
Both sequencing and flow cytometry data are available.
}
\keyword{datasets}
