% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_wkde.R
\name{wkde}
\alias{wkde}
\title{Weighted kernel density estimation, optimized from mixtools package}
\usage{
wkde(x, u = x, w, bw = bw.nrd0(x))
}
\arguments{
\item{x}{The observations}

\item{u}{The point at which to evaluate the density}

\item{w}{the weights, same length as x}

\item{bw}{the bandwith}
}
\value{
densities of the same length as x
}
\description{
Weighted kernel density estimation, optimized from mixtools package
}
