% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_wkdePrep.R
\name{wkdePrep}
\alias{wkdePrep}
\title{Prepare matrix from eighted kernel density estimation,
optimized from mixtools package}
\usage{
wkdePrep(x, u = x, bw)
}
\arguments{
\item{x}{The observations}

\item{u}{The point at which to evaluate the density}

\item{bw}{the bandwith}
}
\value{
densities of the same length as x
}
\description{
Prepare matrix from eighted kernel density estimation,
optimized from mixtools package
}
