% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_calcWeights.R
\name{calcWeights}
\alias{calcWeights}
\title{Obtain weights as posterior probabilities to calculate the consensus null}
\usage{
calcWeights(logDensPerm, fdr, logPriorProbs)
}
\arguments{
\item{logDensPerm}{A matrix with B rows of logged density estimates of the B
permutation distributions}

\item{fdr}{A vector of local false discovery rates
along the support of the kernel}

\item{logPriorProbs}{A vector of length B with logged prior probabilities of
resampling instances}
}
\value{
A vector of weights of length B
}
\description{
Obtain weights as posterior probabilities to calculate the consensus null
}
