% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_calcWeights.R
\name{calcWeights}
\alias{calcWeights}
\title{Obtain weights as posterior probabilities to calculate the consensus null}
\usage{
calcWeights(logDensPerm, weightStrat, fdr, zIndR)
}
\arguments{
\item{weightStrat}{A character string, either "LH" for likelihoods based
on the central region, or "LHw" for weighted likelihoods}

\item{fdr}{A vector of local false discovery rates
along the support of the kernel}

\item{zIndR}{An indicator vector indicating the z-values
within the central region}

\item{densPerm}{A matrix with B rows of density estimates of the B
permutation distributions}

\item{zIndObs}{An indicator vector of the observed z-values
in the support of the kernel}
}
\value{
A vector of weights of length B
}
\description{
Obtain weights as posterior probabilities to calculate the consensus null
}
