% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_estNormal.R
\name{estNormal}
\alias{estNormal}
\title{Fast estimation of mean and standard deviation of a normal disitrbution, optionally with weights}
\usage{
estNormal(y, x, w = NULL, p, B)
}
\arguments{
\item{y}{vector of observations}

\item{x}{design matrix}

\item{w}{optional weight vector}

\item{p}{an integer, number of observations}

\item{B}{an integer, the number of bootstraps}
}
\value{
A vector of length 2 with mean and standard deviation
}
\description{
Fast estimation of mean and standard deviation of a normal disitrbution, optionally with weights
}
