% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_estP0.R
\name{estP0}
\alias{estP0}
\title{Estimate the fraction of true null hypotheses.}
\usage{
estP0(zValObs, nullDensCum, zSeq, z0quantRange, smooth.df)
}
\arguments{
\item{zValObs}{A vector of observed z-values}

\item{nullDensCum}{A smoothed cumulative distribution function of the observed z-values}

\item{zSeq}{The support of the kernel smoother}

\item{z0quantRange}{a number of quantiles between 0 and 0.5}

\item{smooth.df}{degrees of freedom for the spline smoother}
}
\value{
The estimated null fraction, the value of the spline evaluated at the first element of z0quantRange
}
\description{
Estimate the fraction of true null hypotheses.
}
\details{
A natural spline is used over a range of intervals.
Based on the qvalue::qvalue() function and
Storey and Tibshirani, 2003
}
