% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_pt.edit.R
\name{pt.edit}
\alias{pt.edit}
\title{A custom function to calculate the distribution function of the t-test
 statistic. For internal use only}
\usage{
pt.edit(q)
}
\description{
A custom function to calculate the distribution function of the t-test
 statistic. For internal use only
}
