% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_testDAA.R
\docType{methods}
\name{testDAA}
\alias{testDAA}
\alias{testDAA,phyloseq-method}
\alias{testDAA,matrix-method}
\title{A function to test for differential absolute abundance on a phyloseq object}
\usage{
testDAA(Y, ...)

\S4method{testDAA}{phyloseq}(Y, groupName, FCname, ...)

\S4method{testDAA}{matrix}(Y, FC, x, S = rowSums(Y), tieBreak = "none",
  ...)
}
\arguments{
\item{Y}{A phyloseq object, or a data matrix with samples in the rows and
OTUs in the columns}

\item{...}{passed on to the reconsi() function}

\item{groupName}{A character string, the name of a variable in physeq
indicating the grouping factor}

\item{FCname}{A character string, the name of a variable in physeq
containing the total cell count}

\item{FC}{a vector of length n with total flow cytometry cell counts}

\item{x}{a grouping factor of length n}

\item{S}{a vector of library sizes. Will be calculated f}

\item{tieBreak}{A character string, specifying the tie break doctrine to use.}
}
\value{
See the reconsi() function
}
\description{
A function to test for differential absolute abundance on a phyloseq object
}
\examples{
#Test for phyloseq object
testVanDePutte = testDAA(Vandeputte, "Health.status", "absCountFrozen",
B = 5e1)
#Test for matrix
library(phyloseq)
testMat = testDAA(as(otu_table(Vandeputte), "matrix"),
get_variable(Vandeputte, "Health.status"),
get_variable(Vandeputte,"absCountFrozen"), B = 5e1)
}
